Fix the full path where there are nested directories with the same name
This commit fixes the infinite loop when there are nested directories with the same name.
The previous EnsurePrefix method would ensure that a given path starts with the given prefix
already, which caused the paths like actual `path/path` paths to be returned as `path`.
This behavior has been causing the tool to loop forever during searches, as it'd keep
searching for the same `path/` path, without going deeper to the nested `path/path`. The change
in the commit basically adds the prefix all the time without caring if it starts with it.