Fix incomplete/absent errgroup context cancellation handling in folder list worker functions
The Goroutine forked from the pathListWork worker function to write into the pathC channel is missing errgroup context cancellation handling to abort waiting and decrement the waitgroup counter, which blocks the pathC, resC and errC channels from being closed/returned resulting in a hung scenario.

The folderListWork worker function also has incomplete errgroup context cancellation handling which does not return immediately if the thread is waiting to read from the pathC channel and it has not been closed. Though granted pathC would be closed with the fix to pathListWork, this changes the logic to match similar functions in other folder worker functions.